requirements.txt
Django==4.1.9
django-jazzmin==2.6.0   
pillow==9.5.0
djangorestframework==3.14.0